First part of a refactoring that would eliminate JsonUtil.basicTypeFromJson()
This was abandoned for now because it turned out not to be necessary. However, I'm saving it on
this branch because I think there are real performance benefits that could be gained by code-
generating exactly the right conversion logic on each attribute, rather than forcing every
conversion through the decision tree in basicTypeFromJson().